refactor: align spec_decode module naming with vllm upstream
